Added @JsonIgnore in Certificate to prevent loop.
The CertificateResource REST service is fetching the Certificate DAO model
and converts it to JSON. With the new relation between Certificate and
OriginalCertificate this creates an almost infinite loop. This is now
prevented.

NOTE: This is a quick fix. It's bad practice to convert a DB model
directly into a service response. But we can live with this atm since the
CertificateResource only is used for test and Fitnesse.